refactor (CustomBookmarkRepositoryImpl): apply @QueryProjection on findUsedTagIdsWithCount

+++ src/main/java/com/meoguri/linkocean/domain/bookmark/service/dto/GetUsedTagWithCountResult.java
@@ -12,5 +12,5 @@ import lombok.RequiredArgsConstructor;
 public final class GetUsedTagWithCountResult {
 
 	private final String tag;
-	private final int count;
+	private final long count;
 }
